requirements.txt
Django==3.2.25
Pillow==9.5.0
requests==2.31.0
python-dotenv==1.0.0 